some small refactors (#2357)
Reviewed-on: https://gitea.com/xorm/xorm/pulls/2357

+++ delete.go
@@ -63,13 +63,12 @@ func (session *Session) delete(beans []interface{}, mustHaveConditions bool) (in
 	table := session.statement.RefTable
 
 	realSQLWriter := builder.NewWriter()
-	deleteSQLWriter := builder.NewWriter()
-	if err := session.statement.WriteDelete(realSQLWriter, deleteSQLWriter, session.engine.nowTime); err != nil {
+	if err := session.statement.WriteDelete(realSQLWriter, session.engine.nowTime); err != nil {
 		return 0, err
 	}
 
-	if session.statement.GetUnscoped() || table == nil || table.DeletedColumn() == nil { // tag "deleted" is disabled
-	} else {
+	// if tag "deleted" is enabled, then set the field as deleted value
+	if !session.statement.GetUnscoped() && table != nil && table.DeletedColumn() != nil {
 		deletedColumn := table.DeletedColumn()
 		_, t, err := session.engine.nowTime(deletedColumn)
 		if err != nil {
